+++ src/automerge.h
@@ -78,7 +78,7 @@ SEXP C_am_save_incremental(SEXP doc_ptr);
 SEXP C_am_load_incremental(SEXP doc_ptr, SEXP data);
 
 // Shared helpers
-AMresult* convert_r_heads_to_amresult(SEXP heads_list, AMresult ***results_out, size_t *n_results);
+AMresult* convert_r_heads_to_amresult(SEXP heads_list, size_t *n_results);
 AMitems* resolve_heads(SEXP heads, AMitems *heads_items_out, AMresult **heads_result_out);
 
 // Object operations (objects.c)
@@ -89,8 +89,8 @@ SEXP C_am_keys(SEXP doc_ptr, SEXP obj_ptr);
 SEXP C_am_length(SEXP doc_ptr, SEXP obj_ptr);
 SEXP C_am_insert(SEXP doc_ptr, SEXP obj_ptr, SEXP pos, SEXP value);
 SEXP C_am_text_splice(SEXP text_ptr, SEXP pos, SEXP del_count, SEXP text);
-SEXP C_am_text_content(SEXP text_ptr);
-SEXP C_am_text_update(SEXP text_ptr, SEXP old_str, SEXP new_str);
+SEXP C_am_text_content(SEXP text_ptr, SEXP heads);
+SEXP C_am_text_update(SEXP text_ptr, SEXP new_str);
 SEXP C_am_values(SEXP doc_ptr, SEXP obj_ptr);
 SEXP C_am_counter_increment(SEXP doc_ptr, SEXP obj_ptr, SEXP key_or_pos, SEXP delta);
 // v1.2 Object operations (objects.c)